URDF support code (in SimpleGUI/Robot), XML support code (in SimpleGUI/XML) and URDF demo in examples along with katana arm data.
This work was contributed by Emma Fletcher, Johnathan Ross and James
Taylor.

Note due to changes in the project since last commit, the CMakeLists.txt
files will need to updated for these to compile.  When attempting to run
the demo, all the files prefixed by katana in the Examples folder must
be manually copied into the build directory at this point.


git-svn-id: svn+ssh://rabbit.cs.gwu.edu/home/robotgames/repo/branch/SimpleGui@190 4610833d-7e64-4530-9b05-e6b80970a362